+++ foxtive-ntex/src/http/extractors/json_body.rs
@@ -4,6 +4,7 @@ use ntex::http::Payload;
 use ntex::util::BytesMut;
 use ntex::web::{FromRequest, HttpRequest};
 use serde::de::DeserializeOwned;
+use crate::error::HttpError;
 
 pub struct JsonBody {
     json: String,
@@ -24,9 +25,9 @@ impl JsonBody {
 }
 
 impl<Err> FromRequest<Err> for JsonBody {
-    type Error = foxtive::Error;
+    type Error = HttpError;
 
-    async fn from_request(_req: &HttpRequest, payload: &mut Payload) -> AppResult<Self> {
+    async fn from_request(_req: &HttpRequest, payload: &mut Payload) -> Result<JsonBody, Self::Error> {
         let mut bytes = BytesMut::new();
         while let Some(item) = ntex::util::stream_recv(payload).await {
             bytes.extend_from_slice(&item?);