+++ foxtive-ntex/CHANGELOG.md
@@ -3,6 +3,7 @@ Foxtive-Ntex changelog file
 
 ### 0.2.1
 * fix(result): 'is_empty' should return false when value is present
+* fix(responder): should return appropriate status code
 
 ### 0.2.0
 * refactor(defs): renamed to 'ext'